Time: 50 ms (37.08%), Space: 44.5 MB (35.85%) - LeetHub

+++ 0264-ugly-number-ii/0264-ugly-number-ii.java
@@ -14,7 +14,7 @@ class Solution {
             if (seen.add(cur * 5)) pq.add(cur * 5);
         }
         
-        return (int) pq.peek().longValue();
+        return (int) pq.peek().intValue();
     }
 }
